+++ src/lib.rs
@@ -11,4 +11,5 @@ pub mod intel_pstate;
 pub(crate) mod drm;
 pub(crate) mod pseudofs;
 pub(crate) mod sysfs;
+pub(crate) mod units;
 pub(crate) mod utils;